refactor: fix md context generator to using interface
